requirements.txt
statsd-telegraf == 3.2.1-1
